Add basic DNS server address for '8.8.8.8' and '8.8.4.4'
This is just the initial test to make sure the approach works. It is
still lacking error checking and the ability to specify the DNS servers
that are to be used.

In this case, I used two of the well-known Google public DNS servers
for testing. More testing still needs to be done, as well as adding
back in the `-Validate` switch.