Add additional error catching and validation
The script now attempts to validate it can connect to the master and
MSI host before continuing if the `Test-NetConnection` cmdlet is
installed (Server 2012+ and Windows 8.1+).

Additional debug messages included for those who want to run the script
in debug mode (set `$DebugPreference = 'Continue'` in the shell).
Previously squashed output now appears as debug messages. This can
support future development on the module.

Cleaned up the syntax to (generally) use the `${<VAR>}` format for
strings in double-quoted strings. Since PowerShell supports this
syntax and it is now needed in several places in this script, it made
sense to change over to that convention.