add time services support
This commit uses the `W32tm` service to sync the clock of the system.
It does require the current time be within a certain range, otherwise
it will complain and ask for manual intervention.

The service is stopped, the new setting are applied, and then a time
sync is forced once the service has been restarted. If this fails,
the script waits several seconds and then retries the action. If it
fails a second time, the script aborts with an error message.