Add DNS services to script
This commit allows the module to set up basic DNS services for use
during the actual installation. It is expected these settings will
later be managed by Puppet through service-specific classes.

This class allows the configuration of IPv4 and/or IPv6 DNS on one
interface selected by `InterfaceAlias` or `InterfaceIndex`. These
can also be overriden on the command line.

By default, the script will also test the new (or existing) values
to ensure validity before continuing. Administrators also have the
option to not have existing values overriden, if that makes more
sense in their environment. Even in this case, however, the validity
of the existing DNS servers will be checked.