make NTP server support optional
With this module, we really don't want to modify the default behavior,
so we are making whether or not to modify the NTP server settings not
only optional, but also configurable via the commandline at execution
time.

This commit makes the setting of NTP servers optional. By default,
the code will not appear in the generated script, unless the
`set_ntp_servers` parameter is set to `True`.

With this commit, we are also removing the `manifests/params.pp`
file, as it is not necessary.