enable commandline control for DNS servers
With this module, we really don't want to modify the default behavior,
so we are making whether or not ot modify the DNS server settings not
only optional, but also configurable via the commandline at execution
time.

By default, the installation script will no longer attempt to set
the DNS servers. This optional ability is enabled by setting the
parameter `set_dns_servers` to `True`. If this parameter is `False`,
the code for doing DNS setup will not be included in the script.