Rename the script to 'pe_install.ps1'
MAJOR CHANGE:
This commit changes the name of the PowerShell script that this module
makes from 'install.ps1' to 'pe_install.ps1'.

This is being done to prevent a duplicate resource declaration with the
official pe_repo module that comes with Puppet Enterprise. It is likely
that in the near future, a Windows install script will be part of Puppet
Enterprise and will be named 'install.ps1'. At that point, it'd be best
to use Puppet Enterprise's script and stop using this module.